requirements.txt
sklearn
pandas
django
djangorestframework
pyzabbix
seaborn
matplotlib
mysqlclient
